+++ lib/domain/dto/product/product_dto.g.dart
@@ -11,8 +11,8 @@ _ProductDto _$ProductDtoFromJson(Map<String, dynamic> json) => _ProductDto(
   description: json['description'] as String,
   barCode: json['bar_code'] as String,
   saleBy: $enumDecodeNullable(_$SaleByEnumMap, json['sale_by']) ?? SaleBy.unit,
-  category: json['category'] as String?,
-  subcategory: json['subcategory'] as String?,
+  categoryId: json['category_id'] as String?,
+  subCategoryId: json['sub_category_id'] as String?,
   createdAt: json['created_at'] == null
       ? null
       : DateTime.parse(json['created_at'] as String),
@@ -27,8 +27,8 @@ Map<String, dynamic> _$ProductDtoToJson(_ProductDto instance) =>
       'description': instance.description,
       'bar_code': instance.barCode,
       'sale_by': _$SaleByEnumMap[instance.saleBy]!,
-      'category': instance.category,
-      'subcategory': instance.subcategory,
+      'category_id': instance.categoryId,
+      'sub_category_id': instance.subCategoryId,
       'created_at': instance.createdAt?.toIso8601String(),
       'updated_at': instance.updatedAt?.toIso8601String(),
     };