+++ lib/domain/models/product/product_model.g.dart
@@ -14,8 +14,8 @@ _ProductModel _$ProductModelFromJson(Map<String, dynamic> json) =>
       barCode: json['bar_code'] as String,
       saleBy:
           $enumDecodeNullable(_$SaleByEnumMap, json['sale_by']) ?? SaleBy.unit,
-      category: json['category'] as String?,
-      subcategory: json['subcategory'] as String?,
+      categoryId: json['category_id'] as String?,
+      subCategoryId: json['sub_category_id'] as String?,
       createdAt: json['created_at'] == null
           ? null
           : DateTime.parse(json['created_at'] as String),
@@ -31,8 +31,8 @@ Map<String, dynamic> _$ProductModelToJson(_ProductModel instance) =>
       'description': instance.description,
       'bar_code': instance.barCode,
       'sale_by': _$SaleByEnumMap[instance.saleBy]!,
-      'category': instance.category,
-      'subcategory': instance.subcategory,
+      'category_id': instance.categoryId,
+      'sub_category_id': instance.subCategoryId,
       'created_at': instance.createdAt?.toIso8601String(),
       'updated_at': instance.updatedAt?.toIso8601String(),
     };