+++ lib/ui/view/cart_shopping/add_product_cart/add_product_cart_view_model.dart
@@ -1,4 +1,7 @@
+import 'package:compras/domain/models/category/category_model.dart';
 import 'package:compras/domain/models/item/item_model.dart';
+import 'package:compras/domain/models/product/product_model.dart';
+import 'package:compras/domain/models/sub_category/sub_category_model.dart';
 
 import '/domain/user_cases/shopping_cart_user_case.dart';
 import '/utils/command.dart';
@@ -11,11 +14,16 @@ class AddProductCartViewModel {
     // load = Command0<void>(_userCase.load)..execute();
     save = Command1<void, ItemModel>(_save);
     update = Command1<void, ItemModel>(_update);
+    findProductByBarCode = Command1<ProductModel?, String>(
+      _userCase.findProductByBarCode,
+    );
   }
 
   // late final Command0<void> load;
-  late final Command1 save;
-  late final Command1 update;
+  late final Command1<void, ItemModel> save;
+  late final Command1<void, ItemModel> update;
+  late final Command1<ProductModel?, String> findProductByBarCode;
+  late final Command1<(CategoryModel, SubCategoryModel), String> getSubCategory;
 
   Future<Result<void>> _save(ItemModel item) async {
     await _userCase.save(item);